requirements.txt
tensorflow>=2.13.0
transformers>=4.30.0
datasets>=2.14.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
gradio==6.10.0
pytest>=7.0.0
